requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt
python-jose
pydantic
passlib[bcrypt]
pydantic[email]